requirements.txt
rocksdb==0.7.0
python-slugify==5.0.2
